[CHORE] refactor ember-data/-debug into @ember-data/store to fix encapsulation
